Refatoração dos testes
- Movidos testes de /tests para /catalogmanager/tests
- Padronização do código dos testes funcionais
